+++ src/gps.h
@@ -1,9 +1,9 @@
 #pragma once
 #include <stm32g4xx_hal.h>
 
-extern float        gps_latitude;
+extern double       gps_latitude;
 extern char         gps_latitude_chr;
-extern float        gps_longitude;
+extern double       gps_longitude;
 extern char         gps_longitude_chr;
 extern unsigned int gps_sv_number;
 extern float        gps_hdop;
